refactor: propagate zone ID u32 to RPC token, provider, server, tests, and docs
Brings over the remaining u64→u32 changes from PR #276 that were missing:
- RPC auth token (token.rs): field offsets 53→49 bytes
- RPC config, provider, server: zone_id u64→u32
- Justfile: hex format %016x→%08x
- bin/tempo-zone: CLI arg u64→u32
- Integration & unit tests: private_rpc.rs, utils.rs, ws.rs
- Docs: rpc.md, tip-xxxx.mdx

+++ crates/rpc/src/config.rs
@@ -13,7 +13,7 @@ pub struct PrivateRpcConfig {
     /// Zone L2 RPC URL used by zone-specific RPC methods that inspect L2 events.
     pub zone_rpc_url: String,
     /// The zone's numeric identifier.
-    pub zone_id: u64,
+    pub zone_id: u32,
     /// The zone's chain ID.
     pub chain_id: u64,
     /// The ZonePortal contract address on L1.
